+++ lib/colorize_extended.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'colorize'
-require_relative 'colorize/extended'